Allow site contents to use entire width
Setting a limit on the width of the site content results in a lot
of unnecessary whitespace on widescreen monitors, and given how
crowded the cistatus page is getting it would be nice to make use
of that space.  These tweaks to the css allow the content to
stretch as necessary.